+++ tomcat/src/main/java/nextstep/jwp/JwpHttpDispatcher.java
@@ -1,9 +1,9 @@
 package nextstep.jwp;
 
+import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.Handler;
 import org.apache.coyote.http11.HttpDispatcher;
 import org.apache.coyote.http11.StatusCode;
-import org.apache.coyote.http11.request.ContentType;
 import org.apache.coyote.http11.request.Http11Request;
 import org.apache.coyote.http11.response.Http11Response;
 import java.io.File;
