+++ tomcat/src/main/java/nextstep/jwp/handler/get/RegisterGetHandler.java
@@ -1,8 +1,8 @@
 package nextstep.jwp.handler.get;
 
+import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.Handler;
 import org.apache.coyote.http11.StatusCode;
-import org.apache.coyote.http11.request.ContentType;
 import org.apache.coyote.http11.request.Http11Request;
 import org.apache.coyote.http11.response.Http11Response;
 import java.io.File;
