+++ tomcat/src/main/java/nextstep/jwp/handler/post/RegisterPostHandler.java
@@ -3,9 +3,9 @@ package nextstep.jwp.handler.post;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.exception.BusinessException;
 import nextstep.jwp.model.User;
+import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.Handler;
 import org.apache.coyote.http11.StatusCode;
-import org.apache.coyote.http11.request.ContentType;
 import org.apache.coyote.http11.request.Http11Request;
 import org.apache.coyote.http11.request.RequestBody;
 import org.apache.coyote.http11.response.Http11Response;
