+++ tomcat/src/main/java/org/apache/coyote/http11/ContentType.java
@@ -1,6 +1,5 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11;
 
-import org.apache.coyote.http11.HttpException;
 import java.util.Arrays;
 
 public enum ContentType {
