+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequestParser.java
@@ -1,5 +1,7 @@
 package org.apache.coyote.http11.request;
 
+import org.apache.coyote.http11.HttpHeader;
+import org.apache.coyote.http11.HttpMethod;
 import java.io.BufferedReader;
 import java.io.IOException;
