+++ tomcat/src/main/java/org/apache/coyote/http11/response/Http11Response.java
@@ -1,8 +1,8 @@
 package org.apache.coyote.http11.response;
 
+import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.Cookie;
 import org.apache.coyote.http11.StatusCode;
-import org.apache.coyote.http11.request.ContentType;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -12,7 +12,7 @@ public class Http11Response {
     private final ContentType contentType;
     private final String responseBody;
     private final Map<String, String> otherHeader = new HashMap<>();
-    
+
     private Cookie cookie;
 
     public Http11Response(final StatusCode statusCode, final ContentType contentType, final String responseBody) {
